dataCollection/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
asgiref
pytz
psycopg2-binary
sqlparse
python-dotenv

